Add files via upload
"GazeGuard" is an eye recognition-based attendance system. It first registers a new user by capturing an image of their eyes using an IP webcam. This image is then stored in a database. When a user wants to mark their attendance, the system captures a new eye image and compares it with the one stored in the database. If a match is found, the system records the attendance in MongoDB along with the current date and timestamp.